// frontend/src/domain/reportFiles/reportFiles.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../api/graphql";

export type CsvToExcelVariables = {
  storeId: string;
  csvData: string;
  filename: string;
  sheetName?: string | null;
};

export type CsvToExcelResult = {
  csvToExcel: ({
  __typename: "PrintReportNode";
} & {
  fileId: string;
}) | ({
  __typename: "PrintReportError";
} & {
  error: ({
  __typename: "FailedToFetchReportData";
} & {
  errors: unknown;
} & {
  description: string;
});
});
};

export const CsvToExcel = {
  query: "query csvToExcel($storeId: String!, $csvData: String!, $filename: String!, $sheetName: String) {\n  csvToExcel(\n    storeId: $storeId\n    csvData: $csvData\n    filename: $filename\n    sheetName: $sheetName\n  ) {\n    ... on PrintReportNode {\n      __typename\n      fileId\n    }\n    ... on PrintReportError {\n      __typename\n      error {\n        __typename\n        description\n        ... on FailedToFetchReportData {\n          errors\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<CsvToExcelResult, CsvToExcelVariables>;